Rename BondPage to MarketPage
